Clean up imports in backstore module
* Import from siblings not parents.
* Move blkdev_size out of backstore. It was being exported from backstore
so that test code could use it, better to move it up to a higher level so
that it can be used more generically.
* No longer made device module in backstore public. It doesn't need to be
public.
