+++ src/engine/strat_engine/backstore/metadata/bda.rs
@@ -27,7 +27,7 @@ use crate::{
             },
             device::SyncAll,
         },
-        DevUuid, PoolUuid,
+        types::{DevUuid, PoolUuid},
     },
     stratis::{ErrorEnum, StratisError, StratisResult},
 };
